Rearrange import and managed regex usage in tests
